+++ addon/components/t-menu.js
@@ -25,7 +25,7 @@ const animationTypeTable = EmberObject.create({
     * `animation` string - Set the menu animation type.
     * `a` string - Shorthand alias for `animation`.
     * `is-open` boolean - Set initial open menu state.
-    * `action` string - The name of your controller/route action to handle an icon click. Returned with 1 parameter `isOpen`, which is a boolean type indicating if the current state is open or closed.
+    * `onclick` closure action - The name of your consuming application's component/controller/route action to handle the transformicon click. Returned with 1 parameter `isOpen`, which is a boolean type indicating if the current state is open or closed.
 
   Available `animation` types:
     * butterfly
@@ -53,7 +53,7 @@ const animationTypeTable = EmberObject.create({
 export default BaseTransformicon.extend({
   layout,
 
-  'aria-label': 'toggle menu',
+  label: 'toggle menu',
 
   classNameBindings: ['animationType', 'isOpen'],
   /**
@@ -66,7 +66,7 @@ export default BaseTransformicon.extend({
   animationType: computed('animation', {
     get() {
       let anim = get(this, 'animation');
-      return animationTypeTable.get(anim) || animationTypeTable.get(defaultAnimation);
+      return get(animationTypeTable, anim) || get(animationTypeTable, defaultAnimation);
     }
   }),
   /**