+++ addon/components/t-remove.js
@@ -24,7 +24,7 @@ const animationTypeTable = EmberObject.create({
     * `animation` string - Set the menu animation type.
     * `a` string - Shorthand alias for `animation`.
     * `is-removed` boolean - Set initial open removed state.
-    * `action` string - The name of your controller/route action to handle an icon click. Returned with 1 parameter `isRemoved`, which is a boolean type indicating if the current state is pending remove.
+    * `onclick` closure action - The name of your consuming application's component/controller/route action to handle the transformicon click. Returned with 1 parameter `isRemoved`, which is a boolean type indicating if the current state is pending remove.
 
   Available `animation` types:
     * check
@@ -51,7 +51,7 @@ const animationTypeTable = EmberObject.create({
 export default BaseTransformicon.extend({
   layout,
 
-  'aria-label': 'remove item',
+  label: 'remove item',
 
   classNames: ['tcon-remove'],
   classNameBindings: ['animationType', 'isRemoved'],
@@ -65,7 +65,7 @@ export default BaseTransformicon.extend({
   animationType: computed('animation', {
     get() {
       let anim = get(this, 'animation');
-      return animationTypeTable.get(anim) || animationTypeTable.get(defaultAnimation);
+      return get(animationTypeTable, anim) || get(animationTypeTable, defaultAnimation);
     }
   }),
   /**