+++ tests/dummy/app/controllers/grid.js
@@ -2,15 +2,15 @@ import Ember from 'ember';
 
 const {
   Controller,
-  set
+  Logger
 } = Ember;
 
 export default Controller.extend({
   isGridOpen: false,
 
   actions: {
-    updateGridControllerProperty(isOpen) {
-      set(this, 'isGridOpen', isOpen);
+    updateGridAction(isOpen) {
+      Logger.info(`updateGridAction was triggered with the following arguments: isOpen=${isOpen}`);
     }
   }
 });